# docs/swagger.yaml
basePath: /api/public/accounting/v1
definitions:
  dto.AdjustmentRequest:
    properties:
      adjustment_date:
        example: "2026-04-25T00:00:00Z"
        type: string
      adjustment_type:
        enum:
        - WRITE_OFF
        - ROUND_OFF
        - MANUAL_ADJUSTMENT
        - FX_REVALUATION
        example: WRITE_OFF
        type: string
      amount_base:
        example: 500
        type: number
      amount_txn:
        example: 500
        type: number
      notes:
        example: Approved by CFO
        maxLength: 2000
        type: string
      open_item_id:
        example: 550e8400-e29b-41d4-a716-446655440180
        type: string
      reason_code:
        example: BAD_DEBT
        maxLength: 50
        type: string
    required:
    - adjustment_date
    - adjustment_type
    - amount_base
    - amount_txn
    - open_item_id
    type: object
  dto.AllocationRequest:
    properties:
      allocation_amount_base:
        example: 10000
        type: number
      allocation_amount_txn:
        example: 10000
        type: number
      allocation_currency_code:
        example: INR
        type: string
      allocation_date:
        example: "2026-04-20T00:00:00Z"
        type: string
      from_open_item_id:
        example: 550e8400-e29b-41d4-a716-446655440180
        type: string
      to_open_item_id:
        example: 550e8400-e29b-41d4-a716-446655440181
        type: string
    required:
    - allocation_amount_base
    - allocation_amount_txn
    - allocation_currency_code
    - allocation_date
    - from_open_item_id
    - to_open_item_id
    type: object
...
  dto.CreateOpenItemRequest:
    properties:
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      branch_id:
        example: 550e8400-e29b-41d4-a716-446655440080
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      control_account_id:
        example: 550e8400-e29b-41d4-a716-446655440070
        type: string
      currency_code:
        example: INR
        type: string
      document_date:
        example: "2026-04-10T00:00:00Z"
        type: string
      document_no:
        example: INV-1024
        maxLength: 80
        type: string
      due_date:
        example: "2026-05-10T00:00:00Z"
        type: string
      exchange_rate:
        example: 1
        minimum: 0
        type: number
      item_side:
        enum:
        - RECEIVABLE
        - PAYABLE
        example: RECEIVABLE
        type: string
      journal_id:
        example: 550e8400-e29b-41d4-a716-446655440110
        type: string
      original_amount_base:
        example: 50000
        type: number
      original_amount_txn:
        example: 50000
        type: number
      party_id:
        example: 550e8400-e29b-41d4-a716-446655440170
        type: string
      party_type:
        enum:
        - CUSTOMER
        - VENDOR
        - PARTY
        - EMPLOYEE
        - OTHER
        example: CUSTOMER
        type: string
      source_document_id:
        example: 550e8400-e29b-41d4-a716-446655440100
        type: string
      source_document_type:
        example: INVOICE
        maxLength: 80
        type: string
      source_module:
        example: BILLING
        maxLength: 50
        type: string
    required:
    - book_id
    - company_id
    - control_account_id
    - currency_code
    - document_date
    - document_no
    - exchange_rate
    - item_side
    - journal_id
    - original_amount_base
    - original_amount_txn
    - party_id
    - party_type
    - source_document_id
    - source_document_type
    - source_module
    type: object
  dto.CreatePostingRequest:
    properties:
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      branch_id:
        example: 550e8400-e29b-41d4-a716-446655440080
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      idempotency_key:
        example: billing:inv:550e8400-e29b-41d4-a716-446655440100:v1
        maxLength: 200
        type: string
      request_payload:
        type: object
      requested_posting_date:
        example: "2026-04-15T00:00:00Z"
        type: string
      source_document_id:
        example: 550e8400-e29b-41d4-a716-446655440100
        type: string
      source_document_type:
        example: INVOICE
        maxLength: 80
        type: string
      source_event_id:
        example: 550e8400-e29b-41d4-a716-446655440101
        type: string
      source_module:
        example: BILLING
        maxLength: 50
        type: string
    required:
    - book_id
    - company_id
    - idempotency_key
    - request_payload
    - requested_posting_date
    - source_document_id
    - source_document_type
    - source_event_id
    - source_module
    type: object
...
  dto.PostJournalRequest:
    properties:
      journal_id:
        example: 550e8400-e29b-41d4-a716-446655440110
        type: string
      posting_date:
        example: "2026-04-16T00:00:00Z"
        type: string
    required:
    - journal_id
    type: object
...
  dto.UpdateOpenItemRequest:
    properties:
      due_date:
        example: "2026-06-10T00:00:00Z"
        type: string
      item_status:
        enum:
        - OPEN
        - PARTIALLY_ALLOCATED
        - SETTLED
        - WRITTEN_OFF
        - CANCELLED
        example: PARTIALLY_ALLOCATED
        type: string
      remarks:
        example: Customer promised partial payment
        maxLength: 2000
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        items:
          type: string
        type: array
      message:
        type: string
      status:
        type: boolean
      status_code:
        type: integer
    type: object
host: localhost:8086
info:
  contact:
    email: support@queuebuster.co
    name: API Support
    url: http://www.queuebuster.co/support
  description: Enterprise accounting module for QueueBuster ERP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QB Accounting Service API
  version: 1.0.0
paths:
...
  /journals/post:
    post:
      consumes:
      - application/json
      description: Post a draft journal to the ledger
      parameters:
      - description: Post journal request
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/dto.PostJournalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Post a journal
      tags:
      - journals
...
  /open-items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Open item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOpenItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create an open item
      tags:
      - open-items
  /open-items/{id}:
    delete:
      parameters:
      - description: Open item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete an open item
      tags:
      - open-items
    get:
      parameters:
      - description: Open item ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get an open item
      tags:
      - open-items
    put:
      consumes:
      - application/json
      parameters:
      - description: Open item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOpenItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an open item
      tags:
      - open-items
  /open-items/adjustments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Adjustment
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdjustmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create an open item adjustment
      tags:
      - open-items
  /open-items/allocate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Allocation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AllocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Allocate between open items
      tags:
      - open-items
  /open-items/book/{book_id}:
    get:
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List open items by book
      tags:
      - open-items
  /open-items/party/{party_id}:
    get:
      parameters:
      - description: Party ID
        in: path
        name: party_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List open items by party
      tags:
      - open-items
  /posting-requests:
    post:
      consumes:
      - application/json
      description: Enqueue a posting request from a source document
      parameters:
      - description: Posting request data
        in: body
        name: posting_request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePostingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a posting request
      tags:
      - posting-requests
  /posting-requests/{id}:
    get:
      description: Get a posting request by ID
      parameters:
      - description: Posting request ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a posting request
      tags:
      - posting-requests
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"